+++ src/main/java/gov/nasa/pds/harvest/job/parser/AutogenParser.java
@@ -3,7 +3,8 @@ package gov.nasa.pds.harvest.job.parser;
 import java.util.Set;
 
 import org.w3c.dom.Document;
-import gov.nasa.pds.harvest.util.xml.XPathUtils;
+
+import gov.nasa.pds.registry.common.util.xml.XPathUtils;
 
 
 /**
