+++ src/main/java/gov/nasa/pds/harvest/job/parser/BundleConfigParser.java
@@ -10,7 +10,7 @@ import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
 
 import gov.nasa.pds.harvest.job.model.BundleCfg;
-import gov.nasa.pds.harvest.util.xml.XmlDomUtils;
+import gov.nasa.pds.registry.common.util.xml.XmlDomUtils;
 
 
 /**
