+++ src/main/java/gov/nasa/pds/harvest/job/parser/DirsParser.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
-import gov.nasa.pds.harvest.util.xml.XmlDomUtils;
+import gov.nasa.pds.registry.common.util.xml.XmlDomUtils;
 
 
 /**
